Refactor co_mha to use common base class
